pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "pytest-texts-score"
description = "Texts content similarity scoring plugin"
version = "0.1.0"
readme = "README.rst"
requires-python = ">=3.10,<4.0.0"
authors = [
    { name = "Patrik Vodila", email = "vodilpat@cvut.fit.cz" },
]
maintainers = [
    { name = "Patrik Vodila", email = "vodilpat@cvut.fit.cz" },
]
license = {file = "LICENSE"}
classifiers = [
    "Framework :: Pytest",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "pytest>=8.4.2",
    "requests (>=2.32.5,<3.0.0)",
    "langchain-openai (==0.3.14)",
]
[project.urls]
Repository = "https://github.com/VodilaPat/pytest-texts-score"
[project.entry-points.pytest11]
texts-score = "pytest_texts_score.plugin"

[tool.yapf]
based_on_style = "google"

# This ensures it matches Pylint's setting.
column_limit = 80